% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{as_tibble_fix}
\alias{as_tibble_fix}
\title{Quick Fix to as_tibble}
\usage{
as_tibble_fix(d)
}
\arguments{
\item{d}{An object to coerce to a tibble.}
}
\value{
a tibble.
}
\description{
Handle empty lists when coercing to tibble. See
\href{https://github.com/tidyverse/tibble/issues/851}{tibble issue 851}.
}
\keyword{internal}
