% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{basic_query}
\alias{basic_query}
\title{Basic Query}
\usage{
basic_query(url)
}
\arguments{
\item{url}{The query URL.}
}
\value{
The parsed JSON string, as a list.
}
\description{
Helper function for CIMIS query handling.
}
\keyword{internal}
