% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{bind_records}
\alias{bind_records}
\title{Bind Records}
\usage{
bind_records(result)
}
\arguments{
\item{result}{CIMIS query results.}
}
\value{
A data frame.
}
\description{
Bind CIMIS records into a single data frame. This function
is used internally
}
\keyword{internal}
