% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.r
\name{cimir-defunct}
\alias{cimir-defunct}
\alias{get_data}
\alias{get_station}
\alias{get_station_spatial_zipcode}
\alias{get_station_zipcode}
\alias{to_datetime}
\alias{data_items}
\title{Defunct functions in package \pkg{cimir}.}
\usage{
get_data(
  targets,
  start.date,
  end.date,
  items,
  measure.unit = c("E", "M"),
  prioritize.SCS = TRUE
)

get_station(station)

get_station_spatial_zipcode(zipcode)

get_station_zipcode(zipcode)

to_datetime(d)

data_items(type = c("Daily", "Hourly"))
}
\description{
The functions listed below were deprecated in the previous
release and are now defunct. Alternative functions with similar
functionality are mentioned. Help pages for defunct functions are
available at \code{help("cimir-defunct")}.
}
\section{\code{get_data}}{

For \code{get_data}, use \code{\link{cimis_data}}.
}

\section{\code{get_station}}{

For \code{get_station}, use \code{\link{cimis_station}}.
}

\section{\code{get_station_spatial_zipcode}}{

For \code{get_station_spatial_zipcode}, use \code{\link{cimis_spatial_zipcode}}.
}

\section{\code{get_station_zipcode}}{

For \code{get_station_zipcode}, use \code{\link{cimis_zipcode}}.
}

\section{\code{to_datetime}}{

For \code{to_datetime}, use \code{\link{cimis_to_datetime}}.
}

\section{\code{data_items}}{

For \code{data_items}, use \code{\link{cimis_items}}.
}

\keyword{internal}
