% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{cimis_flags}
\alias{cimis_flags}
\title{CIMIS Data Flags}
\usage{
cimis_flags(type = c("Severe", "Informative"), period = "Current")
}
\arguments{
\item{type}{The type of data flag, i.e. \code{"Severe"} or
\code{"Informative"}.}

\item{period}{The Time period that data was collected, i.e.,
"Current" or "Former" (pre-1995).}
}
\value{
a dataframe of data flags.
}
\description{
List CIMIS data quality control flags.
}
\examples{
cimis_flags()
cimis_flags("Informative")
cimis_flags("Severe", period = "Former")

}
\seealso{
\href{https://cimis.water.ca.gov/Resources.aspx}{CIMIS Data Overview - Quality Control}
}
