% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{cimis_items}
\alias{cimis_items}
\title{CIMIS Data Items}
\usage{
cimis_items(type = c("Daily", "Hourly"))
}
\arguments{
\item{type}{The type of data item, i.e. \code{"Daily"} or \code{"Hourly"}.}
}
\value{
a dataframe of data items.
}
\description{
List CIMIS data items.
}
\examples{
cimis_items()

}
