% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{cimis_station}
\alias{cimis_station}
\alias{cimis_spatial_zipcode}
\alias{cimis_zipcode}
\title{Query CIMIS Station Metadata}
\usage{
cimis_station(station)

cimis_spatial_zipcode(zipcode)

cimis_zipcode(zipcode)
}
\arguments{
\item{station}{The station ID. If missing, metadata for all stations
is returned.}

\item{zipcode}{The (spatial) zip code. If missing, metadata for all
stations is returned.}
}
\value{
A \code{tibble} object.
}
\description{
Query CIMIS station metadata.
}
\examples{
if(is_key_set()) {
  cimis_station()
  cimis_zipcode()
  cimis_spatial_zipcode()
}
}
