% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\docType{data}
\name{dataitems}
\alias{dataitems}
\title{dataitems}
\format{
An object of class \code{data.frame} with 43 rows and 4 columns.
}
\usage{
dataitems
}
\description{
A tibble of data items and their names, classes, and providers.
}
\keyword{internal}
