% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{get_station}
\alias{get_station}
\alias{get_station_spatial_zipcode}
\alias{get_station_zipcode}
\title{Get CIMIS Station Data}
\usage{
get_station(station)

get_station_spatial_zipcode(zipcode)

get_station_zipcode(zipcode)
}
\arguments{
\item{station}{The station ID. If missing, metadata for all stations
is returned.}

\item{zipcode}{The (spatial) zip code. If missing, metadata for all
stations is returned.}
}
\value{
A \code{tibble} object.
}
\description{
Get CIMIS station metadata.
}
