% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{to_datetime}
\alias{to_datetime}
\title{To Datetime}
\usage{
to_datetime(d)
}
\arguments{
\item{d}{A data frame of CIMIS data results.}
}
\value{
The data frame, with a new \code{"Datetime"} column replacing
the \code{"Date"} and \code{"Hour"} columns.
}
\description{
Collapse The Date and Hour columns to a single DateTime Column.
}
